Move tests to the package itself
