+++ packages/inertia-vue/tests/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    "extends": [
+        "plugin:cypress/recommended"
+    ]
+}
