+++ packages/inertia-vue/tests/app/Pages/Home.vue
