+++ packages/inertia-vue/tests/app/Pages/Links.vue
