+++ packages/inertia-vue/tests/app/Pages/LinksTarget.vue
