+++ packages/inertia-vue/tests/app/inertia.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <title>Inertia Vue - Testing Environment</title>
+        <script>
+            window.initialPage = '{{ placeholder }}'
+        </script>
+        <script src="app.js" defer></script>
+    </head>
+    <body>
+        <div id="app"></div>
+    </body>
+</html>
