+++ packages/inertia-vue/tests/app/server.js
@@ -10,18 +10,18 @@ app.get(/.*\.js$/, (req, res) => res.sendFile(path.resolve(__dirname, '../tmp/',
 
 // Our actual 'app routes'
 app.get('/', (req, res) => render(req, res, {
-    component: 'Home',
-    props: {
-        example: 'FooBar',
-    }
+  component: 'Home',
+  props: {
+    example: 'FooBar',
+  },
 }))
 
 app.get('/links', (req, res) => render(req, res, {
-    component: 'Links'
+  component: 'Links',
 }))
 
 app.get('/links-target-1', (req, res) => render(req, res, {
-    component: 'LinksTarget'
+  component: 'LinksTarget',
 }))
 
 app.listen(13714)
