+++ packages/inertia-vue/tests/cypress/fixtures/example.json
