+++ packages/inertia-vue/tests/cypress/integration/links.test.js
