+++ packages/inertia-vue/tests/cypress/integration/pages.test.js
