+++ packages/inertia-vue/tests/cypress/plugins/index.js
