+++ packages/inertia-vue/tests/cypress/support/index.js
